Text.Pandoc.Util.Filter.Styling

Documentation

styling :: Block -> IO [Block]

Block-Styling

Special cases captured:

#col turns a div into a floating-div for multiple columns

CodeBlock gets attributes data-trim and data-noescape automatically

.fragment and .frame work properly on divs

inlineStyling :: Inline -> Inline

Inline-Styling

Special cases captured:

.fragment and .frame work properly on spans

.vspace inside span adds a vertical space with height=xxx

.hspace inside span adds a horizontal space with width=xxx